x-monitor/requirements.txt
playwright>=1.40.0
pyyaml>=6.0
openai>=1.0.0
